API

Card

PlayingCards.Suit — Type

Suit

Subtypes are used for each card suit (all of which have aliases):

Club (alias ♣)

Spade (alias ♠)

Heart (alias ♡)

Diamond (alias ♢)

PlayingCards.Card — Type

Card{S <: Suit, R <: Rank}

A playing card. Can be constructed with

Card(suit, rank), or by convenience constructors. For example:

2♢ (equivalent to Card(2, Diamond()))

A♡ (equivalent to Card(1, Heart()))

A 10-suit can be constructed with one of two constructors:

10♣ (equivalent to Card(10, Club()))

or

T♠ (equivalent to Card(10, Spade()))

PlayingCards.suit — Function

suit(::Card)

The card suit (e.g., Heart, Club).

PlayingCards.rank — Function

rank(::Card)

The card rank (e.g., Ace, Rank).

PlayingCards.high_value — Function

high_value(::Card)
high_value(::Rank)

The high rank value. For example:

Rank(1) -> 14 (use low_value for the low Ace value.)

Rank(5) -> 5

PlayingCards.low_value — Function

low_value(::Card)
low_value(::Rank)

The low rank value. For example:

Rank(1) -> 1 (use high_value for the high Ace value.)

Rank(5) -> 5

PlayingCards.color — Function

color(::Card)
color(::Suit)

A Symbol (:red, or :black) indicating the color of the suit or card.

Auxiliary methods

PlayingCards.full_deck — Function

full_deck

A vector of a cards containing a full deck

PlayingCards.ranks — Function

ranks

A Tuple of ranks 1:13.

PlayingCards.suits — Function

suits

A Tuple of all suits

Deck

PlayingCards.Deck — Type

Deck

Deck of cards (backed by a Vector{Card})

PlayingCards.ordered_deck — Function

ordered_deck

An ordered Deck of cards.

Base.pop! — Function

pop!(deck::Deck, n::Int = 1)

Remove n cards from the deck.

Random.shuffle! — Function

shuffle!

Shuffle the deck! shuffle! uses Random.randperm to shuffle the deck.