API

Card

PlayingCards.Suit — Type

Suit

Encode a suit as a 2-bit value (low bits of a UInt8):

0 = ♣ (clubs)

1 = ♢ (diamonds)

2 = ♡ (hearts)

3 = ♠ (spades)

Suits have global constant bindings: ♣, ♢, ♡, ♠.

PlayingCards.Card — Type

Card

Encode a playing card as a 6-bit integer (low bits of a UInt8):

low bits represent rank from 0 to 15

high bits represent suit (♣, ♢, ♡ or ♠)

Ranks are assigned as follows:

numbered cards (2 to 10) have rank equal to their number

jacks, queens and kings have ranks 11, 12 and 13

there are low and high aces with ranks 1 and 14

there are low and high jokers with ranks 0 and 15

This allows any of the standard orderings of cards ranks to be achieved simply by choosing which aces to use.

There are a total of 64 possible card values with this scheme, represented by UInt8 values 0x00 through 0x3f.

PlayingCards.suit — Function

suit(::Card)

The suit of a card

PlayingCards.rank — Function

rank(::Card)

The rank of a card

PlayingCards.high_value — Function

high_value(::Card)
high_value(::Rank)

The high rank value. For example:

Rank(1) -> 14 (use low_value for the low Ace value.)

Rank(5) -> 5

PlayingCards.low_value — Function

low_value(::Card)
low_value(::Rank)

The low rank value. For example:

Rank(1) -> 1 (use high_value for the high Ace value.)

Rank(5) -> 5

PlayingCards.color — Function

color(::Card)

A Symbol (:red, or :black) indicating the color of the suit or card.

Auxiliary methods

PlayingCards.full_deck — Function

full_deck

A vector of a cards containing a full deck

PlayingCards.ranks — Function

ranks

A Tuple of ranks 1:13.

PlayingCards.suits — Function

suits

A Tuple of all suits

Deck

PlayingCards.Deck — Type

Deck

Deck of cards (backed by a Vector{Card})

PlayingCards.ordered_deck — Function

ordered_deck

An ordered Deck of cards.

Base.pop! — Function

pop!(deck::Deck, n::Int = 1)
pop!(deck::Deck, card::Card)

Remove n cards from the deck. or Remove card from the deck.

Shuffling

Random.shuffle! — Function

shuffle!(deck::Deck)

Shuffle the deck. shuffle! uses Random.randperm to shuffle the deck.

Random.shuffle — Function

shuffle(deck::Deck)

Returns a shuffled deck, without modifying the input deck.

PlayingCards.riffle! — Function

riffle!(deck::Deck)

Performs a riffle shuffle on the elements of cards using the Gilbert-Shannon-Reed model.

PlayingCards.riffle — Function

riffle(cards::Vector)
riffle(deck::Deck)

Performs a riffle shuffle permutation on a deep copy of cards using the Gilbert–Shannon–Reeds model. The original Vector of cards is unchanged.

PlayingCards.cut — Function

cut(cards, idx)
cut(cards)

Uses cut! to cut a deep copy of the deck cards leaving the original cards unchanged.

PlayingCards.cut! — Function

cut!(cards, idx = binom_rv(length(cards), 0.5))

Cuts the deck of cards by moving cards 1 through idx (inclusive) to the back of the deck so the first card now was formerly the one at position idx+1.

Cuts the deck at a random location. If the deck has n cards, then the cut location is given by the binomial random variable B(n,1/2).