API

Card

PlayingCards.Suit — Type

Suit

Subtypes are used for each card suit (all of which have aliases):

Club (alias ♣)

Spade (alias ♠)

Heart (alias ♡)

Diamond (alias ♢)

PlayingCards.Rank — Type

Rank

The card rank, subtypes are used for each card rank including

Ace

King

Queen

Jack

NumberCard{N} where 2 ≤ N ≤ 10

PlayingCards.Card — Type

Card{R <: Rank, S <: Suit}

A playing card. Can be constructed with

Card(rank, suit), or by convenience constructors. For example:

2♢ (equivalent to Card(NumberCard(2), Diamond()))

A♡ (equivalent to Card(Ace, Heart()))

A 10-suit can be constructed with one of two constructors:

10♣ (equivalent to Card(NumberCard(10), Club()))

or

T♠ (equivalent to Card(NumberCard(10), Spade()))

PlayingCards.suit — Function

suit(::Card)

The card suit (e.g., Heart, Club).

PlayingCards.rank — Function

rank(::Card)

The card rank (e.g., Ace, Jack, NumberCard{N}).

PlayingCards.rank_type — Function

rank_type(::Card)

The type of the rank.

PlayingCards.high_value — Function

high_value(::Card)
high_value(::Rank)

The rank high_value. For example:

Ace -> 14 (takes high highvalue, use [`lowvalue`](@ref) for low high_value.)

Jack -> 11

NumberCard{N} -> N

PlayingCards.low_value — Function

low_value(::Card)
low_value(::Rank)

The low highvalue of the rank (same as `highvalueexcept forAcefor whichlow_value(Card{Ace}) = 1`.

PlayingCards.color — Function

color(::Card)
color(::Suit)

A Symbol (:red, or :black) indicating the color of the suit or card.

Auxiliary methods

PlayingCards.full_deck — Function

full_deck

A vector of a cards containing a full deck

PlayingCards.ranks — Function

ranks

A Tuple of all ranks.

PlayingCards.suits — Function

suits

A Tuple of all suits

Deck

PlayingCards.Deck — Type

Deck

Deck of cards (backed by a Vector{Card})

PlayingCards.ordered_deck — Function

ordered_deck

An ordered Deck of cards.

Base.pop! — Function

pop!(deck::Deck, n::Int)

Remove n cards from the deck.

Random.shuffle! — Function

shuffle!

Shuffle the deck!